+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 ### Changed
 
 - 🚸(backend) use unaccented full name for user search #1637
+- 🌐(backend) internationalize demo #1644
 
 ## [4.1.0] - 2025-12-09